requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
cloudinary==1.39.1
Django==5.0.3
django-cloudinary-storage==0.3.0
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==24.1
djangorestframework==3.15.0
idna==3.6
pillow==10.2.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
urllib3==2.2.1
whitenoise==6.6.0
